数据流模板删除入参类型修改

+++ src/main/java/com/hydata/intelligence/platform/controller/DataStreamModelController.java
@@ -48,7 +48,7 @@ public class DataStreamModelController {
 		}
 	 
 	 @RequestMapping(value = "/delete" ,method = RequestMethod.GET)
-	    public JSONObject delete(int id){
+	    public JSONObject delete(Long id){
 		 JSONObject params = new JSONObject();
 			params.put("id", id);
 			JSONObject result = CheckParams.checkParams(params);
